DataOffload: Separate data offload transformations into sub-package
